+++ projects/ngx-mat-lib/src/lib/mat-form-field-required/index.ts
@@ -1,0 +1,12 @@
+import {NgModule} from "@angular/core";
+import {MatFormFieldRequiredDirective} from "./mat-form-field-required.directive";
+
+export * from "./mat-form-field-required.directive";
+
+@NgModule({
+    imports: [],
+    exports: [MatFormFieldRequiredDirective],
+    declarations: [MatFormFieldRequiredDirective]
+})
+export class MatFormFieldRequiredModule {
+}