+++ projects/ngx-mat-lib/src/lib/xmat-models/xmat-alert-types.model.ts
@@ -1,0 +1,8 @@
+export enum XmatAlertTypes {
+    success,
+    error,
+    question,
+    info,
+    warning,
+    disabled
+}