+++ projects/ngx-mat-lib/src/lib/xmat-models/xmat-constants-labels.model.ts
@@ -1,0 +1,22 @@
+export interface XmatConstantsLabels {
+    accept: string;
+    cancel: string;
+    changedMind: string;
+    close: string;
+    confirm: string;
+    days: {
+        full: string[];
+        short: string[];
+    };
+    errorTitle: string;
+    genericError: string;
+    leave: string;
+    months: {
+        full: string[];
+        short: string[];
+    };
+    proceed: string;
+    stay: string;
+    understood: string;
+    warningTitle: string;
+}