+++ projects/ngx-mat-lib/src/lib/xmat-models/xmat-desc-value.model.ts
@@ -1,4 +1,5 @@
-export interface XmatDescValuePair<T = any> {
-    desc: string;
+export interface XmatSelect<T = any> {
+    description: string;
     value: T;
 }
+