+++ src/app/services/xmat-test-consts.service.ts
@@ -2,7 +2,7 @@ import {Injectable} from "@angular/core";
 import {
     XmatConstantsService,
     XmatDynamicRestVerbsRef,
-    XmatDescValuePair,
+    XmatSelect,
     XmatGenericObject
 } from "ngx-mat-lib";