+++ src/environments/environment.prod.ts
@@ -1,5 +1,5 @@
 
 export const xmatEnvironment = {
-    appVersion: "7.2.7",
+    appVersion: "7.2.9",
     production: !0
 };